requirements.txt
aiohttp>=3.9.0
web3>=6.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
